Crate llvmint

Bindings to (almost) all LLVM intrinsics.

Intrinsics are categorised into modules by the architecture on which they are supported (however, see Platform support for a caveat), with certain intrinsics available at the top level. These are raw bindings and absolutely zero error checking is performed.

Naming

The intrinsics are named entirely for their internal LLVM names, with the llvm prefix stripped, . replaced by _ and, if it exists, a platform prefix replaced by being placed in a module of that name. For example, llvm.adjust.trampoline becomes ::adjust_trampoline and llvm.x86.addcarry.u32 becomes ::x86::addcarry_u32.

Many of these intrinsics have corresponding intrinsics exposed by GCC/Clang in C/C++, these names are listed in gcc_names.

Platform support

An intrinsic being available in a certain module (or at the top level) does not guarantee that it is supported on all CPUs of that architecture (resp. all CPUs), e.g. the x86::avx512_... intrinsics are only supported on the very latest x86-64 CPUs, not on older x86 processors.

Using an intrinsic in a configuration that is not supported will likely cause LLVM assertions or general badness along those lines.

Modules

AMDGPU
LLVM intrinsics for the AMDGPU architecture.

aarch64
LLVM intrinsics for the aarch64 architecture.

arm
LLVM intrinsics for the arm architecture.

cuda
LLVM intrinsics for the cuda architecture.

gcc_names
Listing of the corresponding name(s) of many GCC intrinsics, for reference/search purposes.

hexagon
LLVM intrinsics for the hexagon architecture.

mips
LLVM intrinsics for the mips architecture.

nvvm
LLVM intrinsics for the nvvm architecture.

ppc
LLVM intrinsics for the ppc architecture.

ptx
LLVM intrinsics for the ptx architecture.

r600
LLVM intrinsics for the r600 architecture.

x86
LLVM intrinsics for the x86 architecture.

xcore
LLVM intrinsics for the xcore architecture.

Functions

adjust_trampoline
The llvm.adjust.trampoline intrinsic; known as __builtin_adjust_trampoline in GCC.

assume
The llvm.assume intrinsic.

clear_cache
The llvm.clear.cache intrinsic.

debugtrap
The llvm.debugtrap intrinsic; known as __builtin_debugtrap in GCC.

donothing
The llvm.donothing intrinsic.

eh_dwarf_cfa
The llvm.eh.dwarf.cfa intrinsic.

eh_return_i32
The llvm.eh.return.i32 intrinsic.

eh_return_i64
The llvm.eh.return.i64 intrinsic.

eh_sjlj_callsite
The llvm.eh.sjlj.callsite intrinsic.

eh_sjlj_functioncontext
The llvm.eh.sjlj.functioncontext intrinsic.

eh_sjlj_longjmp
The llvm.eh.sjlj.longjmp intrinsic.

eh_sjlj_lsda
The llvm.eh.sjlj.lsda intrinsic.

eh_sjlj_setjmp
The llvm.eh.sjlj.setjmp intrinsic.

eh_typeid_for
The llvm.eh.typeid.for intrinsic.

eh_unwind_init
The llvm.eh.unwind.init intrinsic; known as __builtin_unwind_init in GCC.

experimental_patchpoint_i64
The llvm.experimental.patchpoint.i64 intrinsic.

experimental_patchpoint_void
The llvm.experimental.patchpoint.void intrinsic.

experimental_stackmap
The llvm.experimental.stackmap intrinsic.

flt_rounds
The llvm.flt.rounds intrinsic; known as __builtin_flt_rounds in GCC.

frameaddress
The llvm.frameaddress intrinsic.

gcread
The llvm.gcread intrinsic.

gcroot
The llvm.gcroot intrinsic.

gcwrite
The llvm.gcwrite intrinsic.

init_trampoline
The llvm.init.trampoline intrinsic; known as __builtin_init_trampoline in GCC.

lifetime_end
The llvm.lifetime.end intrinsic.

lifetime_start
The llvm.lifetime.start intrinsic.

longjmp
The llvm.longjmp intrinsic.

pcmarker
The llvm.pcmarker intrinsic.

prefetch
The llvm.prefetch intrinsic.

readcyclecounter
The llvm.readcyclecounter intrinsic.

returnaddress
The llvm.returnaddress intrinsic.

setjmp
The llvm.setjmp intrinsic.

siglongjmp
The llvm.siglongjmp intrinsic.

sigsetjmp
The llvm.sigsetjmp intrinsic.

stackprotector
The llvm.stackprotector intrinsic.

stackprotectorcheck
The llvm.stackprotectorcheck intrinsic.

stackrestore
The llvm.stackrestore intrinsic; known as __builtin_stack_restore in GCC.

stacksave
The llvm.stacksave intrinsic; known as __builtin_stack_save in GCC.

trap
The llvm.trap intrinsic; known as __builtin_trap in GCC.

vacopy
The llvm.vacopy intrinsic.

vaend
The llvm.vaend intrinsic.

var_annotation
The llvm.var.annotation intrinsic.

vastart
The llvm.vastart intrinsic.