Crate llvmint

Bindings to (almost) all LLVM intrinsics.

Intrinsics are categorised into modules by the architecture on which they are supported (however, see Platform support for a caveat), with certain intrinsics available at the top level. These are raw bindings and absolutely zero error checking is performed.

Naming

The intrinsics are named entirely for their internal LLVM names, with the llvm prefix stripped, . replaced by _ and, if it exists, a platform prefix replaced by being placed in a module of that name. For example, llvm.adjust.trampoline becomes ::adjust_trampoline and llvm.x86.addcarry.u32 becomes ::x86::addcarry_u32.

Many of these intrinsics have corresponding intrinsics exposed by GCC/Clang in C/C++, these names are listed in gcc_names.

Platform support

An intrinsic being available in a certain module (or at the top level) does not guarantee that it is supported on all CPUs of that architecture (resp. all CPUs), e.g. the x86::avx512_... intrinsics are only supported on the very latest x86-64 CPUs, not on older x86 processors.

Using an intrinsic in a configuration that is not supported will likely cause LLVM assertions or general badness along those lines.

Modules

AMDGPU
LLVM intrinsics for the AMDGPU architecture.

aarch64
LLVM intrinsics for the aarch64 architecture.

arm
LLVM intrinsics for the arm architecture.

cuda
LLVM intrinsics for the cuda architecture.

gcc_names
Listing of the corresponding name(s) of many GCC intrinsics, for reference/search purposes.

hexagon
LLVM intrinsics for the hexagon architecture.

mips
LLVM intrinsics for the mips architecture.

nvvm
LLVM intrinsics for the nvvm architecture.

ppc
LLVM intrinsics for the ppc architecture.

ptx
LLVM intrinsics for the ptx architecture.

x86
LLVM intrinsics for the x86 architecture.

xcore
LLVM intrinsics for the xcore architecture.

Functions

adjust_trampoline
The llvm.adjust.trampoline intrinsic; known as __builtin_adjust_trampoline in GCC.

annotation_i16
The llvm.annotation.i16 intrinsic.

annotation_i32
The llvm.annotation.i32 intrinsic.

annotation_i64
The llvm.annotation.i64 intrinsic.

annotation_i8
The llvm.annotation.i8 intrinsic.

annotation_v16i8
The llvm.annotation.v16i8 intrinsic.

annotation_v2i64
The llvm.annotation.v2i64 intrinsic.

annotation_v4i32
The llvm.annotation.v4i32 intrinsic.

annotation_v8i16
The llvm.annotation.v8i16 intrinsic.

assume
The llvm.assume intrinsic.

bswap_i16
The llvm.bswap.i16 intrinsic.

bswap_i32
The llvm.bswap.i32 intrinsic.

bswap_i64
The llvm.bswap.i64 intrinsic.

bswap_i8
The llvm.bswap.i8 intrinsic.

bswap_v16i8
The llvm.bswap.v16i8 intrinsic.

bswap_v2i64
The llvm.bswap.v2i64 intrinsic.

bswap_v4i32
The llvm.bswap.v4i32 intrinsic.

bswap_v8i16
The llvm.bswap.v8i16 intrinsic.

ceil_f32
The llvm.ceil.f32 intrinsic.

ceil_f64
The llvm.ceil.f64 intrinsic.

ceil_v2f64
The llvm.ceil.v2f64 intrinsic.

ceil_v4f32
The llvm.ceil.v4f32 intrinsic.

clear_cache
The llvm.clear_cache intrinsic.

convert_from_fp16_f32
The llvm.convert.from.fp16.f32 intrinsic.

convert_from_fp16_f64
The llvm.convert.from.fp16.f64 intrinsic.

convert_from_fp16_v2f64
The llvm.convert.from.fp16.v2f64 intrinsic.

convert_from_fp16_v4f32
The llvm.convert.from.fp16.v4f32 intrinsic.

convert_to_fp16_f32
The llvm.convert.to.fp16.f32 intrinsic.

convert_to_fp16_f64
The llvm.convert.to.fp16.f64 intrinsic.

convert_to_fp16_v2f64
The llvm.convert.to.fp16.v2f64 intrinsic.

convert_to_fp16_v4f32
The llvm.convert.to.fp16.v4f32 intrinsic.

convertff_f32_f32
The llvm.convertff.f32.f32 intrinsic.

convertff_f32_f64
The llvm.convertff.f32.f64 intrinsic.

convertff_f32_v2f64
The llvm.convertff.f32.v2f64 intrinsic.

convertff_f32_v4f32
The llvm.convertff.f32.v4f32 intrinsic.

convertff_f64_f32
The llvm.convertff.f64.f32 intrinsic.

convertff_f64_f64
The llvm.convertff.f64.f64 intrinsic.

convertff_f64_v2f64
The llvm.convertff.f64.v2f64 intrinsic.

convertff_f64_v4f32
The llvm.convertff.f64.v4f32 intrinsic.

convertff_v2f64_f32
The llvm.convertff.v2f64.f32 intrinsic.

convertff_v2f64_f64
The llvm.convertff.v2f64.f64 intrinsic.

convertff_v2f64_v2f64
The llvm.convertff.v2f64.v2f64 intrinsic.

convertff_v2f64_v4f32
The llvm.convertff.v2f64.v4f32 intrinsic.

convertff_v4f32_f32
The llvm.convertff.v4f32.f32 intrinsic.

convertff_v4f32_f64
The llvm.convertff.v4f32.f64 intrinsic.

convertff_v4f32_v2f64
The llvm.convertff.v4f32.v2f64 intrinsic.

convertff_v4f32_v4f32
The llvm.convertff.v4f32.v4f32 intrinsic.

convertfsi_f32_i16
The llvm.convertfsi.f32.i16 intrinsic.

convertfsi_f32_i32
The llvm.convertfsi.f32.i32 intrinsic.

convertfsi_f32_i64
The llvm.convertfsi.f32.i64 intrinsic.

convertfsi_f32_i8
The llvm.convertfsi.f32.i8 intrinsic.

convertfsi_f32_v16i8
The llvm.convertfsi.f32.v16i8 intrinsic.

convertfsi_f32_v2i64
The llvm.convertfsi.f32.v2i64 intrinsic.

convertfsi_f32_v4i32
The llvm.convertfsi.f32.v4i32 intrinsic.

convertfsi_f32_v8i16
The llvm.convertfsi.f32.v8i16 intrinsic.

convertfsi_f64_i16
The llvm.convertfsi.f64.i16 intrinsic.

convertfsi_f64_i32
The llvm.convertfsi.f64.i32 intrinsic.

convertfsi_f64_i64
The llvm.convertfsi.f64.i64 intrinsic.

convertfsi_f64_i8
The llvm.convertfsi.f64.i8 intrinsic.

convertfsi_f64_v16i8
The llvm.convertfsi.f64.v16i8 intrinsic.

convertfsi_f64_v2i64
The llvm.convertfsi.f64.v2i64 intrinsic.

convertfsi_f64_v4i32
The llvm.convertfsi.f64.v4i32 intrinsic.

convertfsi_f64_v8i16
The llvm.convertfsi.f64.v8i16 intrinsic.

convertfsi_v2f64_i16
The llvm.convertfsi.v2f64.i16 intrinsic.

convertfsi_v2f64_i32
The llvm.convertfsi.v2f64.i32 intrinsic.

convertfsi_v2f64_i64
The llvm.convertfsi.v2f64.i64 intrinsic.

convertfsi_v2f64_i8
The llvm.convertfsi.v2f64.i8 intrinsic.

convertfsi_v2f64_v16i8
The llvm.convertfsi.v2f64.v16i8 intrinsic.

convertfsi_v2f64_v2i64
The llvm.convertfsi.v2f64.v2i64 intrinsic.

convertfsi_v2f64_v4i32
The llvm.convertfsi.v2f64.v4i32 intrinsic.

convertfsi_v2f64_v8i16
The llvm.convertfsi.v2f64.v8i16 intrinsic.

convertfsi_v4f32_i16
The llvm.convertfsi.v4f32.i16 intrinsic.

convertfsi_v4f32_i32
The llvm.convertfsi.v4f32.i32 intrinsic.

convertfsi_v4f32_i64
The llvm.convertfsi.v4f32.i64 intrinsic.

convertfsi_v4f32_i8
The llvm.convertfsi.v4f32.i8 intrinsic.

convertfsi_v4f32_v16i8
The llvm.convertfsi.v4f32.v16i8 intrinsic.

convertfsi_v4f32_v2i64
The llvm.convertfsi.v4f32.v2i64 intrinsic.

convertfsi_v4f32_v4i32
The llvm.convertfsi.v4f32.v4i32 intrinsic.

convertfsi_v4f32_v8i16
The llvm.convertfsi.v4f32.v8i16 intrinsic.

convertfui_f32_i16
The llvm.convertfui.f32.i16 intrinsic.

convertfui_f32_i32
The llvm.convertfui.f32.i32 intrinsic.

convertfui_f32_i64
The llvm.convertfui.f32.i64 intrinsic.

convertfui_f32_i8
The llvm.convertfui.f32.i8 intrinsic.

convertfui_f32_v16i8
The llvm.convertfui.f32.v16i8 intrinsic.

convertfui_f32_v2i64
The llvm.convertfui.f32.v2i64 intrinsic.

convertfui_f32_v4i32
The llvm.convertfui.f32.v4i32 intrinsic.

convertfui_f32_v8i16
The llvm.convertfui.f32.v8i16 intrinsic.

convertfui_f64_i16
The llvm.convertfui.f64.i16 intrinsic.

convertfui_f64_i32
The llvm.convertfui.f64.i32 intrinsic.

convertfui_f64_i64
The llvm.convertfui.f64.i64 intrinsic.

convertfui_f64_i8
The llvm.convertfui.f64.i8 intrinsic.

convertfui_f64_v16i8
The llvm.convertfui.f64.v16i8 intrinsic.

convertfui_f64_v2i64
The llvm.convertfui.f64.v2i64 intrinsic.

convertfui_f64_v4i32
The llvm.convertfui.f64.v4i32 intrinsic.

convertfui_f64_v8i16
The llvm.convertfui.f64.v8i16 intrinsic.

convertfui_v2f64_i16
The llvm.convertfui.v2f64.i16 intrinsic.

convertfui_v2f64_i32
The llvm.convertfui.v2f64.i32 intrinsic.

convertfui_v2f64_i64
The llvm.convertfui.v2f64.i64 intrinsic.

convertfui_v2f64_i8
The llvm.convertfui.v2f64.i8 intrinsic.

convertfui_v2f64_v16i8
The llvm.convertfui.v2f64.v16i8 intrinsic.

convertfui_v2f64_v2i64
The llvm.convertfui.v2f64.v2i64 intrinsic.

convertfui_v2f64_v4i32
The llvm.convertfui.v2f64.v4i32 intrinsic.

convertfui_v2f64_v8i16
The llvm.convertfui.v2f64.v8i16 intrinsic.

convertfui_v4f32_i16
The llvm.convertfui.v4f32.i16 intrinsic.

convertfui_v4f32_i32
The llvm.convertfui.v4f32.i32 intrinsic.

convertfui_v4f32_i64
The llvm.convertfui.v4f32.i64 intrinsic.

convertfui_v4f32_i8
The llvm.convertfui.v4f32.i8 intrinsic.

convertfui_v4f32_v16i8
The llvm.convertfui.v4f32.v16i8 intrinsic.

convertfui_v4f32_v2i64
The llvm.convertfui.v4f32.v2i64 intrinsic.

convertfui_v4f32_v4i32
The llvm.convertfui.v4f32.v4i32 intrinsic.

convertfui_v4f32_v8i16
The llvm.convertfui.v4f32.v8i16 intrinsic.

convertsif_i16_f32
The llvm.convertsif.i16.f32 intrinsic.

convertsif_i16_f64
The llvm.convertsif.i16.f64 intrinsic.

convertsif_i16_v2f64
The llvm.convertsif.i16.v2f64 intrinsic.

convertsif_i16_v4f32
The llvm.convertsif.i16.v4f32 intrinsic.

convertsif_i32_f32
The llvm.convertsif.i32.f32 intrinsic.

convertsif_i32_f64
The llvm.convertsif.i32.f64 intrinsic.

convertsif_i32_v2f64
The llvm.convertsif.i32.v2f64 intrinsic.

convertsif_i32_v4f32
The llvm.convertsif.i32.v4f32 intrinsic.

convertsif_i64_f32
The llvm.convertsif.i64.f32 intrinsic.

convertsif_i64_f64
The llvm.convertsif.i64.f64 intrinsic.

convertsif_i64_v2f64
The llvm.convertsif.i64.v2f64 intrinsic.

convertsif_i64_v4f32
The llvm.convertsif.i64.v4f32 intrinsic.

convertsif_i8_f32
The llvm.convertsif.i8.f32 intrinsic.

convertsif_i8_f64
The llvm.convertsif.i8.f64 intrinsic.

convertsif_i8_v2f64
The llvm.convertsif.i8.v2f64 intrinsic.

convertsif_i8_v4f32
The llvm.convertsif.i8.v4f32 intrinsic.

convertsif_v16i8_f32
The llvm.convertsif.v16i8.f32 intrinsic.

convertsif_v16i8_f64
The llvm.convertsif.v16i8.f64 intrinsic.

convertsif_v16i8_v2f64
The llvm.convertsif.v16i8.v2f64 intrinsic.

convertsif_v16i8_v4f32
The llvm.convertsif.v16i8.v4f32 intrinsic.

convertsif_v2i64_f32
The llvm.convertsif.v2i64.f32 intrinsic.

convertsif_v2i64_f64
The llvm.convertsif.v2i64.f64 intrinsic.

convertsif_v2i64_v2f64
The llvm.convertsif.v2i64.v2f64 intrinsic.

convertsif_v2i64_v4f32
The llvm.convertsif.v2i64.v4f32 intrinsic.

convertsif_v4i32_f32
The llvm.convertsif.v4i32.f32 intrinsic.

convertsif_v4i32_f64
The llvm.convertsif.v4i32.f64 intrinsic.

convertsif_v4i32_v2f64
The llvm.convertsif.v4i32.v2f64 intrinsic.

convertsif_v4i32_v4f32
The llvm.convertsif.v4i32.v4f32 intrinsic.

convertsif_v8i16_f32
The llvm.convertsif.v8i16.f32 intrinsic.

convertsif_v8i16_f64
The llvm.convertsif.v8i16.f64 intrinsic.

convertsif_v8i16_v2f64
The llvm.convertsif.v8i16.v2f64 intrinsic.

convertsif_v8i16_v4f32
The llvm.convertsif.v8i16.v4f32 intrinsic.

convertss_i16_i16
The llvm.convertss.i16.i16 intrinsic.

convertss_i16_i32
The llvm.convertss.i16.i32 intrinsic.

convertss_i16_i64
The llvm.convertss.i16.i64 intrinsic.

convertss_i16_i8
The llvm.convertss.i16.i8 intrinsic.

convertss_i16_v16i8
The llvm.convertss.i16.v16i8 intrinsic.

convertss_i16_v2i64
The llvm.convertss.i16.v2i64 intrinsic.

convertss_i16_v4i32
The llvm.convertss.i16.v4i32 intrinsic.

convertss_i16_v8i16
The llvm.convertss.i16.v8i16 intrinsic.

convertss_i32_i16
The llvm.convertss.i32.i16 intrinsic.

convertss_i32_i32
The llvm.convertss.i32.i32 intrinsic.

convertss_i32_i64
The llvm.convertss.i32.i64 intrinsic.

convertss_i32_i8
The llvm.convertss.i32.i8 intrinsic.

convertss_i32_v16i8
The llvm.convertss.i32.v16i8 intrinsic.

convertss_i32_v2i64
The llvm.convertss.i32.v2i64 intrinsic.

convertss_i32_v4i32
The llvm.convertss.i32.v4i32 intrinsic.

convertss_i32_v8i16
The llvm.convertss.i32.v8i16 intrinsic.

convertss_i64_i16
The llvm.convertss.i64.i16 intrinsic.

convertss_i64_i32
The llvm.convertss.i64.i32 intrinsic.

convertss_i64_i64
The llvm.convertss.i64.i64 intrinsic.

convertss_i64_i8
The llvm.convertss.i64.i8 intrinsic.

convertss_i64_v16i8
The llvm.convertss.i64.v16i8 intrinsic.

convertss_i64_v2i64
The llvm.convertss.i64.v2i64 intrinsic.

convertss_i64_v4i32
The llvm.convertss.i64.v4i32 intrinsic.

convertss_i64_v8i16
The llvm.convertss.i64.v8i16 intrinsic.

convertss_i8_i16
The llvm.convertss.i8.i16 intrinsic.

convertss_i8_i32
The llvm.convertss.i8.i32 intrinsic.

convertss_i8_i64
The llvm.convertss.i8.i64 intrinsic.

convertss_i8_i8
The llvm.convertss.i8.i8 intrinsic.

convertss_i8_v16i8
The llvm.convertss.i8.v16i8 intrinsic.

convertss_i8_v2i64
The llvm.convertss.i8.v2i64 intrinsic.

convertss_i8_v4i32
The llvm.convertss.i8.v4i32 intrinsic.

convertss_i8_v8i16
The llvm.convertss.i8.v8i16 intrinsic.

convertss_v16i8_i16
The llvm.convertss.v16i8.i16 intrinsic.

convertss_v16i8_i32
The llvm.convertss.v16i8.i32 intrinsic.

convertss_v16i8_i64
The llvm.convertss.v16i8.i64 intrinsic.

convertss_v16i8_i8
The llvm.convertss.v16i8.i8 intrinsic.

convertss_v16i8_v16i8
The llvm.convertss.v16i8.v16i8 intrinsic.

convertss_v16i8_v2i64
The llvm.convertss.v16i8.v2i64 intrinsic.

convertss_v16i8_v4i32
The llvm.convertss.v16i8.v4i32 intrinsic.

convertss_v16i8_v8i16
The llvm.convertss.v16i8.v8i16 intrinsic.

convertss_v2i64_i16
The llvm.convertss.v2i64.i16 intrinsic.

convertss_v2i64_i32
The llvm.convertss.v2i64.i32 intrinsic.

convertss_v2i64_i64
The llvm.convertss.v2i64.i64 intrinsic.

convertss_v2i64_i8
The llvm.convertss.v2i64.i8 intrinsic.

convertss_v2i64_v16i8
The llvm.convertss.v2i64.v16i8 intrinsic.

convertss_v2i64_v2i64
The llvm.convertss.v2i64.v2i64 intrinsic.

convertss_v2i64_v4i32
The llvm.convertss.v2i64.v4i32 intrinsic.

convertss_v2i64_v8i16
The llvm.convertss.v2i64.v8i16 intrinsic.

convertss_v4i32_i16
The llvm.convertss.v4i32.i16 intrinsic.

convertss_v4i32_i32
The llvm.convertss.v4i32.i32 intrinsic.

convertss_v4i32_i64
The llvm.convertss.v4i32.i64 intrinsic.

convertss_v4i32_i8
The llvm.convertss.v4i32.i8 intrinsic.

convertss_v4i32_v16i8
The llvm.convertss.v4i32.v16i8 intrinsic.

convertss_v4i32_v2i64
The llvm.convertss.v4i32.v2i64 intrinsic.

convertss_v4i32_v4i32
The llvm.convertss.v4i32.v4i32 intrinsic.

convertss_v4i32_v8i16
The llvm.convertss.v4i32.v8i16 intrinsic.

convertss_v8i16_i16
The llvm.convertss.v8i16.i16 intrinsic.

convertss_v8i16_i32
The llvm.convertss.v8i16.i32 intrinsic.

convertss_v8i16_i64
The llvm.convertss.v8i16.i64 intrinsic.

convertss_v8i16_i8
The llvm.convertss.v8i16.i8 intrinsic.

convertss_v8i16_v16i8
The llvm.convertss.v8i16.v16i8 intrinsic.

convertss_v8i16_v2i64
The llvm.convertss.v8i16.v2i64 intrinsic.

convertss_v8i16_v4i32
The llvm.convertss.v8i16.v4i32 intrinsic.

convertss_v8i16_v8i16
The llvm.convertss.v8i16.v8i16 intrinsic.

convertsu_i16_i16
The llvm.convertsu.i16.i16 intrinsic.

convertsu_i16_i32
The llvm.convertsu.i16.i32 intrinsic.

convertsu_i16_i64
The llvm.convertsu.i16.i64 intrinsic.

convertsu_i16_i8
The llvm.convertsu.i16.i8 intrinsic.

convertsu_i16_v16i8
The llvm.convertsu.i16.v16i8 intrinsic.

convertsu_i16_v2i64
The llvm.convertsu.i16.v2i64 intrinsic.

convertsu_i16_v4i32
The llvm.convertsu.i16.v4i32 intrinsic.

convertsu_i16_v8i16
The llvm.convertsu.i16.v8i16 intrinsic.

convertsu_i32_i16
The llvm.convertsu.i32.i16 intrinsic.

convertsu_i32_i32
The llvm.convertsu.i32.i32 intrinsic.

convertsu_i32_i64
The llvm.convertsu.i32.i64 intrinsic.

convertsu_i32_i8
The llvm.convertsu.i32.i8 intrinsic.

convertsu_i32_v16i8
The llvm.convertsu.i32.v16i8 intrinsic.

convertsu_i32_v2i64
The llvm.convertsu.i32.v2i64 intrinsic.

convertsu_i32_v4i32
The llvm.convertsu.i32.v4i32 intrinsic.

convertsu_i32_v8i16
The llvm.convertsu.i32.v8i16 intrinsic.

convertsu_i64_i16
The llvm.convertsu.i64.i16 intrinsic.

convertsu_i64_i32
The llvm.convertsu.i64.i32 intrinsic.

convertsu_i64_i64
The llvm.convertsu.i64.i64 intrinsic.

convertsu_i64_i8
The llvm.convertsu.i64.i8 intrinsic.

convertsu_i64_v16i8
The llvm.convertsu.i64.v16i8 intrinsic.

convertsu_i64_v2i64
The llvm.convertsu.i64.v2i64 intrinsic.

convertsu_i64_v4i32
The llvm.convertsu.i64.v4i32 intrinsic.

convertsu_i64_v8i16
The llvm.convertsu.i64.v8i16 intrinsic.

convertsu_i8_i16
The llvm.convertsu.i8.i16 intrinsic.

convertsu_i8_i32
The llvm.convertsu.i8.i32 intrinsic.

convertsu_i8_i64
The llvm.convertsu.i8.i64 intrinsic.

convertsu_i8_i8
The llvm.convertsu.i8.i8 intrinsic.

convertsu_i8_v16i8
The llvm.convertsu.i8.v16i8 intrinsic.

convertsu_i8_v2i64
The llvm.convertsu.i8.v2i64 intrinsic.

convertsu_i8_v4i32
The llvm.convertsu.i8.v4i32 intrinsic.

convertsu_i8_v8i16
The llvm.convertsu.i8.v8i16 intrinsic.

convertsu_v16i8_i16
The llvm.convertsu.v16i8.i16 intrinsic.

convertsu_v16i8_i32
The llvm.convertsu.v16i8.i32 intrinsic.

convertsu_v16i8_i64
The llvm.convertsu.v16i8.i64 intrinsic.

convertsu_v16i8_i8
The llvm.convertsu.v16i8.i8 intrinsic.

convertsu_v16i8_v16i8
The llvm.convertsu.v16i8.v16i8 intrinsic.

convertsu_v16i8_v2i64
The llvm.convertsu.v16i8.v2i64 intrinsic.

convertsu_v16i8_v4i32
The llvm.convertsu.v16i8.v4i32 intrinsic.

convertsu_v16i8_v8i16
The llvm.convertsu.v16i8.v8i16 intrinsic.

convertsu_v2i64_i16
The llvm.convertsu.v2i64.i16 intrinsic.

convertsu_v2i64_i32
The llvm.convertsu.v2i64.i32 intrinsic.

convertsu_v2i64_i64
The llvm.convertsu.v2i64.i64 intrinsic.

convertsu_v2i64_i8
The llvm.convertsu.v2i64.i8 intrinsic.

convertsu_v2i64_v16i8
The llvm.convertsu.v2i64.v16i8 intrinsic.

convertsu_v2i64_v2i64
The llvm.convertsu.v2i64.v2i64 intrinsic.

convertsu_v2i64_v4i32
The llvm.convertsu.v2i64.v4i32 intrinsic.

convertsu_v2i64_v8i16
The llvm.convertsu.v2i64.v8i16 intrinsic.

convertsu_v4i32_i16
The llvm.convertsu.v4i32.i16 intrinsic.

convertsu_v4i32_i32
The llvm.convertsu.v4i32.i32 intrinsic.

convertsu_v4i32_i64
The llvm.convertsu.v4i32.i64 intrinsic.

convertsu_v4i32_i8
The llvm.convertsu.v4i32.i8 intrinsic.

convertsu_v4i32_v16i8
The llvm.convertsu.v4i32.v16i8 intrinsic.

convertsu_v4i32_v2i64
The llvm.convertsu.v4i32.v2i64 intrinsic.

convertsu_v4i32_v4i32
The llvm.convertsu.v4i32.v4i32 intrinsic.

convertsu_v4i32_v8i16
The llvm.convertsu.v4i32.v8i16 intrinsic.

convertsu_v8i16_i16
The llvm.convertsu.v8i16.i16 intrinsic.

convertsu_v8i16_i32
The llvm.convertsu.v8i16.i32 intrinsic.

convertsu_v8i16_i64
The llvm.convertsu.v8i16.i64 intrinsic.

convertsu_v8i16_i8
The llvm.convertsu.v8i16.i8 intrinsic.

convertsu_v8i16_v16i8
The llvm.convertsu.v8i16.v16i8 intrinsic.

convertsu_v8i16_v2i64
The llvm.convertsu.v8i16.v2i64 intrinsic.

convertsu_v8i16_v4i32
The llvm.convertsu.v8i16.v4i32 intrinsic.

convertsu_v8i16_v8i16
The llvm.convertsu.v8i16.v8i16 intrinsic.

convertuif_i16_f32
The llvm.convertuif.i16.f32 intrinsic.

convertuif_i16_f64
The llvm.convertuif.i16.f64 intrinsic.

convertuif_i16_v2f64
The llvm.convertuif.i16.v2f64 intrinsic.

convertuif_i16_v4f32
The llvm.convertuif.i16.v4f32 intrinsic.

convertuif_i32_f32
The llvm.convertuif.i32.f32 intrinsic.

convertuif_i32_f64
The llvm.convertuif.i32.f64 intrinsic.

convertuif_i32_v2f64
The llvm.convertuif.i32.v2f64 intrinsic.

convertuif_i32_v4f32
The llvm.convertuif.i32.v4f32 intrinsic.

convertuif_i64_f32
The llvm.convertuif.i64.f32 intrinsic.

convertuif_i64_f64
The llvm.convertuif.i64.f64 intrinsic.

convertuif_i64_v2f64
The llvm.convertuif.i64.v2f64 intrinsic.

convertuif_i64_v4f32
The llvm.convertuif.i64.v4f32 intrinsic.

convertuif_i8_f32
The llvm.convertuif.i8.f32 intrinsic.

convertuif_i8_f64
The llvm.convertuif.i8.f64 intrinsic.

convertuif_i8_v2f64
The llvm.convertuif.i8.v2f64 intrinsic.

convertuif_i8_v4f32
The llvm.convertuif.i8.v4f32 intrinsic.

convertuif_v16i8_f32
The llvm.convertuif.v16i8.f32 intrinsic.

convertuif_v16i8_f64
The llvm.convertuif.v16i8.f64 intrinsic.

convertuif_v16i8_v2f64
The llvm.convertuif.v16i8.v2f64 intrinsic.

convertuif_v16i8_v4f32
The llvm.convertuif.v16i8.v4f32 intrinsic.

convertuif_v2i64_f32
The llvm.convertuif.v2i64.f32 intrinsic.

convertuif_v2i64_f64
The llvm.convertuif.v2i64.f64 intrinsic.

convertuif_v2i64_v2f64
The llvm.convertuif.v2i64.v2f64 intrinsic.

convertuif_v2i64_v4f32
The llvm.convertuif.v2i64.v4f32 intrinsic.

convertuif_v4i32_f32
The llvm.convertuif.v4i32.f32 intrinsic.

convertuif_v4i32_f64
The llvm.convertuif.v4i32.f64 intrinsic.

convertuif_v4i32_v2f64
The llvm.convertuif.v4i32.v2f64 intrinsic.

convertuif_v4i32_v4f32
The llvm.convertuif.v4i32.v4f32 intrinsic.

convertuif_v8i16_f32
The llvm.convertuif.v8i16.f32 intrinsic.

convertuif_v8i16_f64
The llvm.convertuif.v8i16.f64 intrinsic.

convertuif_v8i16_v2f64
The llvm.convertuif.v8i16.v2f64 intrinsic.

convertuif_v8i16_v4f32
The llvm.convertuif.v8i16.v4f32 intrinsic.

convertus_i16_i16
The llvm.convertus.i16.i16 intrinsic.

convertus_i16_i32
The llvm.convertus.i16.i32 intrinsic.

convertus_i16_i64
The llvm.convertus.i16.i64 intrinsic.

convertus_i16_i8
The llvm.convertus.i16.i8 intrinsic.

convertus_i16_v16i8
The llvm.convertus.i16.v16i8 intrinsic.

convertus_i16_v2i64
The llvm.convertus.i16.v2i64 intrinsic.

convertus_i16_v4i32
The llvm.convertus.i16.v4i32 intrinsic.

convertus_i16_v8i16
The llvm.convertus.i16.v8i16 intrinsic.

convertus_i32_i16
The llvm.convertus.i32.i16 intrinsic.

convertus_i32_i32
The llvm.convertus.i32.i32 intrinsic.

convertus_i32_i64
The llvm.convertus.i32.i64 intrinsic.

convertus_i32_i8
The llvm.convertus.i32.i8 intrinsic.

convertus_i32_v16i8
The llvm.convertus.i32.v16i8 intrinsic.

convertus_i32_v2i64
The llvm.convertus.i32.v2i64 intrinsic.

convertus_i32_v4i32
The llvm.convertus.i32.v4i32 intrinsic.

convertus_i32_v8i16
The llvm.convertus.i32.v8i16 intrinsic.

convertus_i64_i16
The llvm.convertus.i64.i16 intrinsic.

convertus_i64_i32
The llvm.convertus.i64.i32 intrinsic.

convertus_i64_i64
The llvm.convertus.i64.i64 intrinsic.

convertus_i64_i8
The llvm.convertus.i64.i8 intrinsic.

convertus_i64_v16i8
The llvm.convertus.i64.v16i8 intrinsic.

convertus_i64_v2i64
The llvm.convertus.i64.v2i64 intrinsic.

convertus_i64_v4i32
The llvm.convertus.i64.v4i32 intrinsic.

convertus_i64_v8i16
The llvm.convertus.i64.v8i16 intrinsic.

convertus_i8_i16
The llvm.convertus.i8.i16 intrinsic.

convertus_i8_i32
The llvm.convertus.i8.i32 intrinsic.

convertus_i8_i64
The llvm.convertus.i8.i64 intrinsic.

convertus_i8_i8
The llvm.convertus.i8.i8 intrinsic.

convertus_i8_v16i8
The llvm.convertus.i8.v16i8 intrinsic.

convertus_i8_v2i64
The llvm.convertus.i8.v2i64 intrinsic.

convertus_i8_v4i32
The llvm.convertus.i8.v4i32 intrinsic.

convertus_i8_v8i16
The llvm.convertus.i8.v8i16 intrinsic.

convertus_v16i8_i16
The llvm.convertus.v16i8.i16 intrinsic.

convertus_v16i8_i32
The llvm.convertus.v16i8.i32 intrinsic.

convertus_v16i8_i64
The llvm.convertus.v16i8.i64 intrinsic.

convertus_v16i8_i8
The llvm.convertus.v16i8.i8 intrinsic.

convertus_v16i8_v16i8
The llvm.convertus.v16i8.v16i8 intrinsic.

convertus_v16i8_v2i64
The llvm.convertus.v16i8.v2i64 intrinsic.

convertus_v16i8_v4i32
The llvm.convertus.v16i8.v4i32 intrinsic.

convertus_v16i8_v8i16
The llvm.convertus.v16i8.v8i16 intrinsic.

convertus_v2i64_i16
The llvm.convertus.v2i64.i16 intrinsic.

convertus_v2i64_i32
The llvm.convertus.v2i64.i32 intrinsic.

convertus_v2i64_i64
The llvm.convertus.v2i64.i64 intrinsic.

convertus_v2i64_i8
The llvm.convertus.v2i64.i8 intrinsic.

convertus_v2i64_v16i8
The llvm.convertus.v2i64.v16i8 intrinsic.

convertus_v2i64_v2i64
The llvm.convertus.v2i64.v2i64 intrinsic.

convertus_v2i64_v4i32
The llvm.convertus.v2i64.v4i32 intrinsic.

convertus_v2i64_v8i16
The llvm.convertus.v2i64.v8i16 intrinsic.

convertus_v4i32_i16
The llvm.convertus.v4i32.i16 intrinsic.

convertus_v4i32_i32
The llvm.convertus.v4i32.i32 intrinsic.

convertus_v4i32_i64
The llvm.convertus.v4i32.i64 intrinsic.

convertus_v4i32_i8
The llvm.convertus.v4i32.i8 intrinsic.

convertus_v4i32_v16i8
The llvm.convertus.v4i32.v16i8 intrinsic.

convertus_v4i32_v2i64
The llvm.convertus.v4i32.v2i64 intrinsic.

convertus_v4i32_v4i32
The llvm.convertus.v4i32.v4i32 intrinsic.

convertus_v4i32_v8i16
The llvm.convertus.v4i32.v8i16 intrinsic.

convertus_v8i16_i16
The llvm.convertus.v8i16.i16 intrinsic.

convertus_v8i16_i32
The llvm.convertus.v8i16.i32 intrinsic.

convertus_v8i16_i64
The llvm.convertus.v8i16.i64 intrinsic.

convertus_v8i16_i8
The llvm.convertus.v8i16.i8 intrinsic.

convertus_v8i16_v16i8
The llvm.convertus.v8i16.v16i8 intrinsic.

convertus_v8i16_v2i64
The llvm.convertus.v8i16.v2i64 intrinsic.

convertus_v8i16_v4i32
The llvm.convertus.v8i16.v4i32 intrinsic.

convertus_v8i16_v8i16
The llvm.convertus.v8i16.v8i16 intrinsic.

convertuu_i16_i16
The llvm.convertuu.i16.i16 intrinsic.

convertuu_i16_i32
The llvm.convertuu.i16.i32 intrinsic.

convertuu_i16_i64
The llvm.convertuu.i16.i64 intrinsic.

convertuu_i16_i8
The llvm.convertuu.i16.i8 intrinsic.

convertuu_i16_v16i8
The llvm.convertuu.i16.v16i8 intrinsic.

convertuu_i16_v2i64
The llvm.convertuu.i16.v2i64 intrinsic.

convertuu_i16_v4i32
The llvm.convertuu.i16.v4i32 intrinsic.

convertuu_i16_v8i16
The llvm.convertuu.i16.v8i16 intrinsic.

convertuu_i32_i16
The llvm.convertuu.i32.i16 intrinsic.

convertuu_i32_i32
The llvm.convertuu.i32.i32 intrinsic.

convertuu_i32_i64
The llvm.convertuu.i32.i64 intrinsic.

convertuu_i32_i8
The llvm.convertuu.i32.i8 intrinsic.

convertuu_i32_v16i8
The llvm.convertuu.i32.v16i8 intrinsic.

convertuu_i32_v2i64
The llvm.convertuu.i32.v2i64 intrinsic.

convertuu_i32_v4i32
The llvm.convertuu.i32.v4i32 intrinsic.

convertuu_i32_v8i16
The llvm.convertuu.i32.v8i16 intrinsic.

convertuu_i64_i16
The llvm.convertuu.i64.i16 intrinsic.

convertuu_i64_i32
The llvm.convertuu.i64.i32 intrinsic.

convertuu_i64_i64
The llvm.convertuu.i64.i64 intrinsic.

convertuu_i64_i8
The llvm.convertuu.i64.i8 intrinsic.

convertuu_i64_v16i8
The llvm.convertuu.i64.v16i8 intrinsic.

convertuu_i64_v2i64
The llvm.convertuu.i64.v2i64 intrinsic.

convertuu_i64_v4i32
The llvm.convertuu.i64.v4i32 intrinsic.

convertuu_i64_v8i16
The llvm.convertuu.i64.v8i16 intrinsic.

convertuu_i8_i16
The llvm.convertuu.i8.i16 intrinsic.

convertuu_i8_i32
The llvm.convertuu.i8.i32 intrinsic.

convertuu_i8_i64
The llvm.convertuu.i8.i64 intrinsic.

convertuu_i8_i8
The llvm.convertuu.i8.i8 intrinsic.

convertuu_i8_v16i8
The llvm.convertuu.i8.v16i8 intrinsic.

convertuu_i8_v2i64
The llvm.convertuu.i8.v2i64 intrinsic.

convertuu_i8_v4i32
The llvm.convertuu.i8.v4i32 intrinsic.

convertuu_i8_v8i16
The llvm.convertuu.i8.v8i16 intrinsic.

convertuu_v16i8_i16
The llvm.convertuu.v16i8.i16 intrinsic.

convertuu_v16i8_i32
The llvm.convertuu.v16i8.i32 intrinsic.

convertuu_v16i8_i64
The llvm.convertuu.v16i8.i64 intrinsic.

convertuu_v16i8_i8
The llvm.convertuu.v16i8.i8 intrinsic.

convertuu_v16i8_v16i8
The llvm.convertuu.v16i8.v16i8 intrinsic.

convertuu_v16i8_v2i64
The llvm.convertuu.v16i8.v2i64 intrinsic.

convertuu_v16i8_v4i32
The llvm.convertuu.v16i8.v4i32 intrinsic.

convertuu_v16i8_v8i16
The llvm.convertuu.v16i8.v8i16 intrinsic.

convertuu_v2i64_i16
The llvm.convertuu.v2i64.i16 intrinsic.

convertuu_v2i64_i32
The llvm.convertuu.v2i64.i32 intrinsic.

convertuu_v2i64_i64
The llvm.convertuu.v2i64.i64 intrinsic.

convertuu_v2i64_i8
The llvm.convertuu.v2i64.i8 intrinsic.

convertuu_v2i64_v16i8
The llvm.convertuu.v2i64.v16i8 intrinsic.

convertuu_v2i64_v2i64
The llvm.convertuu.v2i64.v2i64 intrinsic.

convertuu_v2i64_v4i32
The llvm.convertuu.v2i64.v4i32 intrinsic.

convertuu_v2i64_v8i16
The llvm.convertuu.v2i64.v8i16 intrinsic.

convertuu_v4i32_i16
The llvm.convertuu.v4i32.i16 intrinsic.

convertuu_v4i32_i32
The llvm.convertuu.v4i32.i32 intrinsic.

convertuu_v4i32_i64
The llvm.convertuu.v4i32.i64 intrinsic.

convertuu_v4i32_i8
The llvm.convertuu.v4i32.i8 intrinsic.

convertuu_v4i32_v16i8
The llvm.convertuu.v4i32.v16i8 intrinsic.

convertuu_v4i32_v2i64
The llvm.convertuu.v4i32.v2i64 intrinsic.

convertuu_v4i32_v4i32
The llvm.convertuu.v4i32.v4i32 intrinsic.

convertuu_v4i32_v8i16
The llvm.convertuu.v4i32.v8i16 intrinsic.

convertuu_v8i16_i16
The llvm.convertuu.v8i16.i16 intrinsic.

convertuu_v8i16_i32
The llvm.convertuu.v8i16.i32 intrinsic.

convertuu_v8i16_i64
The llvm.convertuu.v8i16.i64 intrinsic.

convertuu_v8i16_i8
The llvm.convertuu.v8i16.i8 intrinsic.

convertuu_v8i16_v16i8
The llvm.convertuu.v8i16.v16i8 intrinsic.

convertuu_v8i16_v2i64
The llvm.convertuu.v8i16.v2i64 intrinsic.

convertuu_v8i16_v4i32
The llvm.convertuu.v8i16.v4i32 intrinsic.

convertuu_v8i16_v8i16
The llvm.convertuu.v8i16.v8i16 intrinsic.

copysign_f32
The llvm.copysign.f32 intrinsic.

copysign_f64
The llvm.copysign.f64 intrinsic.

copysign_v2f64
The llvm.copysign.v2f64 intrinsic.

copysign_v4f32
The llvm.copysign.v4f32 intrinsic.

cos_f32
The llvm.cos.f32 intrinsic.

cos_f64
The llvm.cos.f64 intrinsic.

cos_v2f64
The llvm.cos.v2f64 intrinsic.

cos_v4f32
The llvm.cos.v4f32 intrinsic.

ctlz_i16
The llvm.ctlz.i16 intrinsic.

ctlz_i32
The llvm.ctlz.i32 intrinsic.

ctlz_i64
The llvm.ctlz.i64 intrinsic.

ctlz_i8
The llvm.ctlz.i8 intrinsic.

ctlz_v16i8
The llvm.ctlz.v16i8 intrinsic.

ctlz_v2i64
The llvm.ctlz.v2i64 intrinsic.

ctlz_v4i32
The llvm.ctlz.v4i32 intrinsic.

ctlz_v8i16
The llvm.ctlz.v8i16 intrinsic.

ctpop_i16
The llvm.ctpop.i16 intrinsic.

ctpop_i32
The llvm.ctpop.i32 intrinsic.

ctpop_i64
The llvm.ctpop.i64 intrinsic.

ctpop_i8
The llvm.ctpop.i8 intrinsic.

ctpop_v16i8
The llvm.ctpop.v16i8 intrinsic.

ctpop_v2i64
The llvm.ctpop.v2i64 intrinsic.

ctpop_v4i32
The llvm.ctpop.v4i32 intrinsic.

ctpop_v8i16
The llvm.ctpop.v8i16 intrinsic.

cttz_i16
The llvm.cttz.i16 intrinsic.

cttz_i32
The llvm.cttz.i32 intrinsic.

cttz_i64
The llvm.cttz.i64 intrinsic.

cttz_i8
The llvm.cttz.i8 intrinsic.

cttz_v16i8
The llvm.cttz.v16i8 intrinsic.

cttz_v2i64
The llvm.cttz.v2i64 intrinsic.

cttz_v4i32
The llvm.cttz.v4i32 intrinsic.

cttz_v8i16
The llvm.cttz.v8i16 intrinsic.

debugtrap
The llvm.debugtrap intrinsic; known as __builtin_debugtrap in GCC.

donothing
The llvm.donothing intrinsic.

eh_dwarf_cfa
The llvm.eh.dwarf.cfa intrinsic.

eh_return_i32
The llvm.eh.return.i32 intrinsic.

eh_return_i64
The llvm.eh.return.i64 intrinsic.

eh_sjlj_callsite
The llvm.eh.sjlj.callsite intrinsic.

eh_sjlj_functioncontext
The llvm.eh.sjlj.functioncontext intrinsic.

eh_sjlj_longjmp
The llvm.eh.sjlj.longjmp intrinsic.

eh_sjlj_lsda
The llvm.eh.sjlj.lsda intrinsic.

eh_sjlj_setjmp
The llvm.eh.sjlj.setjmp intrinsic.

eh_typeid_for
The llvm.eh.typeid.for intrinsic.

eh_unwind_init
The llvm.eh.unwind.init intrinsic; known as __builtin_unwind_init in GCC.

exp2_f32
The llvm.exp2.f32 intrinsic.

exp2_f64
The llvm.exp2.f64 intrinsic.

exp2_v2f64
The llvm.exp2.v2f64 intrinsic.

exp2_v4f32
The llvm.exp2.v4f32 intrinsic.

exp_f32
The llvm.exp.f32 intrinsic.

exp_f64
The llvm.exp.f64 intrinsic.

exp_v2f64
The llvm.exp.v2f64 intrinsic.

exp_v4f32
The llvm.exp.v4f32 intrinsic.

expect_i16
The llvm.expect.i16 intrinsic.

expect_i32
The llvm.expect.i32 intrinsic.

expect_i64
The llvm.expect.i64 intrinsic.

expect_i8
The llvm.expect.i8 intrinsic.

expect_v16i8
The llvm.expect.v16i8 intrinsic.

expect_v2i64
The llvm.expect.v2i64 intrinsic.

expect_v4i32
The llvm.expect.v4i32 intrinsic.

expect_v8i16
The llvm.expect.v8i16 intrinsic.

experimental_patchpoint_i64
The llvm.experimental.patchpoint.i64 intrinsic.

experimental_patchpoint_void
The llvm.experimental.patchpoint.void intrinsic.

experimental_stackmap
The llvm.experimental.stackmap intrinsic.

fabs_f32
The llvm.fabs.f32 intrinsic.

fabs_f64
The llvm.fabs.f64 intrinsic.

fabs_v2f64
The llvm.fabs.v2f64 intrinsic.

fabs_v4f32
The llvm.fabs.v4f32 intrinsic.

floor_f32
The llvm.floor.f32 intrinsic.

floor_f64
The llvm.floor.f64 intrinsic.

floor_v2f64
The llvm.floor.v2f64 intrinsic.

floor_v4f32
The llvm.floor.v4f32 intrinsic.

flt_rounds
The llvm.flt.rounds intrinsic; known as __builtin_flt_rounds in GCC.

fma_f32
The llvm.fma.f32 intrinsic.

fma_f64
The llvm.fma.f64 intrinsic.

fma_v2f64
The llvm.fma.v2f64 intrinsic.

fma_v4f32
The llvm.fma.v4f32 intrinsic.

fmuladd_f32
The llvm.fmuladd.f32 intrinsic.

fmuladd_f64
The llvm.fmuladd.f64 intrinsic.

fmuladd_v2f64
The llvm.fmuladd.v2f64 intrinsic.

fmuladd_v4f32
The llvm.fmuladd.v4f32 intrinsic.

frameaddress
The llvm.frameaddress intrinsic.

gcread
The llvm.gcread intrinsic.

gcroot
The llvm.gcroot intrinsic.

gcwrite
The llvm.gcwrite intrinsic.

init_trampoline
The llvm.init.trampoline intrinsic; known as __builtin_init_trampoline in GCC.

lifetime_end
The llvm.lifetime.end intrinsic.

lifetime_start
The llvm.lifetime.start intrinsic.

log10_f32
The llvm.log10.f32 intrinsic.

log10_f64
The llvm.log10.f64 intrinsic.

log10_v2f64
The llvm.log10.v2f64 intrinsic.

log10_v4f32
The llvm.log10.v4f32 intrinsic.

log2_f32
The llvm.log2.f32 intrinsic.

log2_f64
The llvm.log2.f64 intrinsic.

log2_v2f64
The llvm.log2.v2f64 intrinsic.

log2_v4f32
The llvm.log2.v4f32 intrinsic.

log_f32
The llvm.log.f32 intrinsic.

log_f64
The llvm.log.f64 intrinsic.

log_v2f64
The llvm.log.v2f64 intrinsic.

log_v4f32
The llvm.log.v4f32 intrinsic.

longjmp
The llvm.longjmp intrinsic.

memcpy_p0i8_p0i8_i16
The llvm.memcpy.p0i8.p0i8.i16 intrinsic.

memcpy_p0i8_p0i8_i32
The llvm.memcpy.p0i8.p0i8.i32 intrinsic.

memcpy_p0i8_p0i8_i64
The llvm.memcpy.p0i8.p0i8.i64 intrinsic.

memcpy_p0i8_p0i8_i8
The llvm.memcpy.p0i8.p0i8.i8 intrinsic.

memcpy_p0i8_p0i8_v16i8
The llvm.memcpy.p0i8.p0i8.v16i8 intrinsic.

memcpy_p0i8_p0i8_v2i64
The llvm.memcpy.p0i8.p0i8.v2i64 intrinsic.

memcpy_p0i8_p0i8_v4i32
The llvm.memcpy.p0i8.p0i8.v4i32 intrinsic.

memcpy_p0i8_p0i8_v8i16
The llvm.memcpy.p0i8.p0i8.v8i16 intrinsic.

memmove_p0i8_p0i8_i16
The llvm.memmove.p0i8.p0i8.i16 intrinsic.

memmove_p0i8_p0i8_i32
The llvm.memmove.p0i8.p0i8.i32 intrinsic.

memmove_p0i8_p0i8_i64
The llvm.memmove.p0i8.p0i8.i64 intrinsic.

memmove_p0i8_p0i8_i8
The llvm.memmove.p0i8.p0i8.i8 intrinsic.

memmove_p0i8_p0i8_v16i8
The llvm.memmove.p0i8.p0i8.v16i8 intrinsic.

memmove_p0i8_p0i8_v2i64
The llvm.memmove.p0i8.p0i8.v2i64 intrinsic.

memmove_p0i8_p0i8_v4i32
The llvm.memmove.p0i8.p0i8.v4i32 intrinsic.

memmove_p0i8_p0i8_v8i16
The llvm.memmove.p0i8.p0i8.v8i16 intrinsic.

memset_p0i8_i16
The llvm.memset.p0i8.i16 intrinsic.

memset_p0i8_i32
The llvm.memset.p0i8.i32 intrinsic.

memset_p0i8_i64
The llvm.memset.p0i8.i64 intrinsic.

memset_p0i8_i8
The llvm.memset.p0i8.i8 intrinsic.

memset_p0i8_v16i8
The llvm.memset.p0i8.v16i8 intrinsic.

memset_p0i8_v2i64
The llvm.memset.p0i8.v2i64 intrinsic.

memset_p0i8_v4i32
The llvm.memset.p0i8.v4i32 intrinsic.

memset_p0i8_v8i16
The llvm.memset.p0i8.v8i16 intrinsic.

nearbyint_f32
The llvm.nearbyint.f32 intrinsic.

nearbyint_f64
The llvm.nearbyint.f64 intrinsic.

nearbyint_v2f64
The llvm.nearbyint.v2f64 intrinsic.

nearbyint_v4f32
The llvm.nearbyint.v4f32 intrinsic.

objectsize_i16_p0i8
The llvm.objectsize.i16.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_i32_p0i8
The llvm.objectsize.i32.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_i64_p0i8
The llvm.objectsize.i64.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_i8_p0i8
The llvm.objectsize.i8.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_v16i8_p0i8
The llvm.objectsize.v16i8.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_v2i64_p0i8
The llvm.objectsize.v2i64.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_v4i32_p0i8
The llvm.objectsize.v4i32.p0i8 intrinsic; known as __builtin_object_size in GCC.

objectsize_v8i16_p0i8
The llvm.objectsize.v8i16.p0i8 intrinsic; known as __builtin_object_size in GCC.

pcmarker
The llvm.pcmarker intrinsic.

pow_f32
The llvm.pow.f32 intrinsic.

pow_f64
The llvm.pow.f64 intrinsic.

pow_v2f64
The llvm.pow.v2f64 intrinsic.

pow_v4f32
The llvm.pow.v4f32 intrinsic.

powi_f32
The llvm.powi.f32 intrinsic.

powi_f64
The llvm.powi.f64 intrinsic.

powi_v2f64
The llvm.powi.v2f64 intrinsic.

powi_v4f32
The llvm.powi.v4f32 intrinsic.

prefetch
The llvm.prefetch intrinsic.

ptr_annotation_p0i16
The llvm.ptr.annotation.p0i16 intrinsic.

ptr_annotation_p0i32
The llvm.ptr.annotation.p0i32 intrinsic.

ptr_annotation_p0i64
The llvm.ptr.annotation.p0i64 intrinsic.

ptr_annotation_p0i8
The llvm.ptr.annotation.p0i8 intrinsic.

ptr_annotation_p0v16i8
The llvm.ptr.annotation.p0v16i8 intrinsic.

ptr_annotation_p0v2i64
The llvm.ptr.annotation.p0v2i64 intrinsic.

ptr_annotation_p0v4i32
The llvm.ptr.annotation.p0v4i32 intrinsic.

ptr_annotation_p0v8i16
The llvm.ptr.annotation.p0v8i16 intrinsic.

readcyclecounter
The llvm.readcyclecounter intrinsic.

returnaddress
The llvm.returnaddress intrinsic.

rint_f32
The llvm.rint.f32 intrinsic.

rint_f64
The llvm.rint.f64 intrinsic.

rint_v2f64
The llvm.rint.v2f64 intrinsic.

rint_v4f32
The llvm.rint.v4f32 intrinsic.

round_f32
The llvm.round.f32 intrinsic.

round_f64
The llvm.round.f64 intrinsic.

round_v2f64
The llvm.round.v2f64 intrinsic.

round_v4f32
The llvm.round.v4f32 intrinsic.

setjmp
The llvm.setjmp intrinsic.

siglongjmp
The llvm.siglongjmp intrinsic.

sigsetjmp
The llvm.sigsetjmp intrinsic.

sin_f32
The llvm.sin.f32 intrinsic.

sin_f64
The llvm.sin.f64 intrinsic.

sin_v2f64
The llvm.sin.v2f64 intrinsic.

sin_v4f32
The llvm.sin.v4f32 intrinsic.

sqrt_f32
The llvm.sqrt.f32 intrinsic.

sqrt_f64
The llvm.sqrt.f64 intrinsic.

sqrt_v2f64
The llvm.sqrt.v2f64 intrinsic.

sqrt_v4f32
The llvm.sqrt.v4f32 intrinsic.

stackprotector
The llvm.stackprotector intrinsic.

stackprotectorcheck
The llvm.stackprotectorcheck intrinsic.

stackrestore
The llvm.stackrestore intrinsic; known as __builtin_stack_restore in GCC.

stacksave
The llvm.stacksave intrinsic; known as __builtin_stack_save in GCC.

trap
The llvm.trap intrinsic; known as __builtin_trap in GCC.

trunc_f32
The llvm.trunc.f32 intrinsic.

trunc_f64
The llvm.trunc.f64 intrinsic.

trunc_v2f64
The llvm.trunc.v2f64 intrinsic.

trunc_v4f32
The llvm.trunc.v4f32 intrinsic.

vacopy
The llvm.va_copy intrinsic.

vaend
The llvm.va_end intrinsic.

var_annotation
The llvm.var.annotation intrinsic.

vastart
The llvm.va_start intrinsic.